services/auth_service/requirements.txt
fastapi==0.109.1
pydantic==2.11.4
pydantic-settings==2.9.1
starlette==0.35.1

uvicorn[standard]==0.22.0
sqlalchemy==1.4.47

requests==2.32.0
httpx==0.27.0
cryptography==44.0.1
pyodbc~=5.2.0
python-dotenv~=1.1.0
passlib[bcrypt]
python-jose[cryptography]
nltk~=3.9.1              # используется в sentiment-анализе
sentiment~=0.0.1         # если кастомная библиотека есть

# поддержка C и сетевых библиотек
cffi~=1.17.1
pycparser~=2.22

# доп. утилиты и совместимость
h11~=0.16.0
httpcore~=1.0.9
sniffio~=1.3.1
anyio~=4.9.0
click~=8.2.0
typing_extensions~=4.13.2
exceptiongroup~=1.3.0
annotated-types~=0.7.0
charset-normalizer~=3.4.2
certifi~=2025.4.26
urllib3~=2.4.0
idna~=3.10
httptools~=0.6.4
websockets~=15.0.1
watchfiles~=1.0.5
greenlet~=3.2.2

# build инструменты
pip~=25.1.1
setuptools~=70.0.0
wheel~=0.41.2
PyYAML~=6.0.2
python-multipart